Refactor command number variable to use unsigned int

+++ src/Client/commands.cpp
@@ -75,7 +75,7 @@ void CommandManager::waitForCommand(UserState& state) {
     std::istringstream iss(line);
     std::string  number_str;
     iss >> number_str;
-    int number = std::stoi(number_str);
+    u_int32_t number = static_cast<uint32_t>(std::stoi(number_str));
     std::cout << "Sleeping for "<< number << " seconds..." << std::endl;
     sleep(number);
     return;
